updated for keithley support
