+++ examples/example_fake_instruments.py
