+++ examples/example_real_instruments.py
@@ -1,0 +1,5 @@
+import spcs_instruments as spcs 
+keithley = spcs.Keithley2400("C:\\Users\\david\\Documents\\spcs_instruments\\templates\\keithley_config_example.toml")
+keithley.measure
+
+pfx-p
