+++ src/spcs_instruments/__init__.py
@@ -1,6 +1,6 @@
 from .instruments import Fake_daq
 from .instruments import SiglentSDS2352XE
-
+from .instruments import Keithley2400
 from .spcs_instruments_utils import Experiment
 
-__all__ = ["Fake_daq","SiglentSDS2352XE", "Experiment"]
+__all__ = ["Fake_daq","SiglentSDS2352XE", "Experiment", "Keithley2400"]
