+++ src/spcs_instruments/instruments/__init__.py
@@ -1,4 +1,4 @@
 from .test_instrument import Fake_daq
 from .siglent_sds_2352xe_driver import SiglentSDS2352XE
-
-__all__ = ["Fake_daq", "SiglentSDS2352XE"]
+from .keithley_2400_driver import Keithley2400
+__all__ = ["Fake_daq", "SiglentSDS2352XE", "Keithley2400"]
